requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==5.0.3
django-extensions==3.2.3
djangorestframework==3.14.0
fonttools==4.49.0
h11==0.14.0
html5lib==1.1
idna==3.6
outcome==1.3.0.post0
pdfkit==1.0.0
pillow==10.2.0
pycparser==2.21
pydyf==0.9.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyphen==0.14.0
PySocks==1.7.1
pytz==2024.1
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
weasyprint==61.2
webencodings==0.5.1
wsproto==1.2.0
zopfli==0.2.3
